feat, refactor: Added error message support if internet is not availabe and refactored code
Added error message support if internet is not availabe. Showing error message as a visual indicator that your internet is not working so the user can take an action.
Basic refactoring code and removed commented code.
Also fixed the blank page when returning from detail activity by not destroying the loader.

Resolves #23, resolves #22